server/requirements.txt
flask
openai
anthropic>=0.42.0
gunicorn
flask-cors
requests
psycopg2-binary
stripe>=7.0.0
PyJWT>=2.8.0
APScheduler>=3.10.0
